+++ ChangeLog.md
@@ -2,6 +2,12 @@
 
 ## [Unreleased]
 
+## [0.1.2.9] - 2020-03-12
+### Added
+- Function `filterAtomsOfModel` to filter atoms of model by the given predicate.
+### Fixed
+- Grouping by residues when converting `Mae` to `Model`.
+
 ## [0.1.2.8] - 2020-03-12
 ### Added
 - `instance Traversable (Sequence mk w)`.